From "Michael Pheasant <mike@pheasant.co.nz>"
Database-list ignores a supplied database name and instead uses
"mysql". It is possible that a mysql userid has permissions to issue
"show databases" but no permission to connect to db "mysql".
database-list gives an error in this scenario.

The following patch overcomes this by using the user-supplied db name
or else "mysql" if nil supplied.