iso8601 fractional second bug fixes
Previously printing and then parsing an iso-timestring would lose
the fractional seconds, which causes errors in postgres (at least).
Db-timestring was changed to call iso-timestring because it was
outputting the same format (as the old iso-timestring), but with
only one the bug related to usecs.

iso-timestring used to output
  XXXX-XX-XX XX:XX:XX,,XXXXXX

What was invalid in this format was the two commas and
because 3 usec would be output as xx,,3 instead of xx,,000003

it now outputs
  XXXX-XX-XXTXX:XX:XX.XXXXXXZ
which should be a valid iso8601 timestamp and is parseable by
parse-timestring and parse-iso-8601-time